+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## `V2.0.0`
+
+- Added predefined variables which can be used too run the shell commands
+  Example :
+  ```js
+  exports.Run_File = "node ${file}" // can use predefined variables
+  ```
 ## `V1.0.0`
 
 - Initial release